Only do surface relative decodings relative to a previously found global position.
Surface relative decoding is ambiguous within approx 45nm, and
receiver range could easily be greater than that, so finding a
position relative to the receiver is uncertain. It's safer to
just wait for a second position message and do a global decoding
in this case to get things started; subsequent messages can be
used relative to the global position safely.